pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Kobina Brandon <kobinabrandon@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pandas = "^2.2.3"
loguru = "^0.7.3"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.7.1"
requests = "^2.32.3"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
